refactor message handling with context provider

+++ frontend/src/contexts/MessageContextProvider.jsx
@@ -1,11 +1,13 @@
 import { createContext, useContext, useState, useEffect } from "react"
 import { useItemIsActiveStatus } from "./ItemIsActiveStatusContextProvider"
+import { useClientContext } from "./ClientContextProvider";
 
 export const MessageContext = createContext();
 
 const MessageContextProvider = ({ children }) => {
     const { itemIsActiveStatus } = useItemIsActiveStatus();
-    
+    const { actualClientData } = useClientContext();
+
     const [messageList, setMessageList] = useState([]);
     const [newOrModifiedMessage, setNewOrModifiedMessage] = useState([]);
     const [originalMessage, setOriginalMessage] = useState("");
@@ -16,7 +18,7 @@ const MessageContextProvider = ({ children }) => {
 
     const messagesFetch = async (url) => {
         const actualUrl = `${url}?isActive=${itemIsActiveStatus}`
-        
+
         try {
             const response = await fetch(actualUrl);
             if (!response.ok) {
@@ -28,49 +30,55 @@ const MessageContextProvider = ({ children }) => {
             console.error("Error fetching messages:", error.message);
         }
     };
-    
+
     useEffect(() => {
         messagesFetch(messageUrl);
     }, [itemIsActiveStatus]);
 
-    
-    async function deleteMessageFetch(removableMessageId) {
-        try {
-            const requestOptions = {
-                method: 'DELETE',
-                headers: { 'Content-Type': 'application/json' }
-            };
-
-            const response = await fetch(`${messageUrl}/${removableMessageId}`, requestOptions);
-            if (response.status === 200) {
-                const newMessageList = messageList.filter(message => message._id !== removableMessageId);
-                setMessageList(newMessageList)
-                console.log('Message delete was successful.');
-            } else {
-                throw new Error("Failed to delete message.")
-            }
-        }
-        catch (error) {
-            console.error("Problem with message delete!", error.message)
-        }
+    const updateMessage = (index, messageId) => {
+        const actualMessage = messageList.find(message => message._id === messageId);
+        actualMessage.message = `${(index + 1)}.) ${actualMessage.message}`;
+        setOriginalMessage(actualMessage.message)
+        setNewOrModifiedMessage({ message: actualMessage.message, _id: messageId })
     };
 
-    async function removeMessageFetch(removableMessageId) {
-        try {
-            const requestOptions = {
-                method: 'PUT',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({ isActive: false })
-            };
-            const response = await fetch(`${messageUrl}/${removableMessageId}`, requestOptions);
-            if (response.status === 200) {
-                const newMessageList = messageList.filter(message => message._id !== removableMessageId);
-                setMessageList(newMessageList)
-                console.log('Message remove was successful');
+    async function deleteMessage(removableMessageId) {
+        if (actualClientData.bossStatus === true) {
+            try {
+                const requestOptions = {
+                    method: 'DELETE',
+                    headers: { 'Content-Type': 'application/json' }
+                };
+
+                const response = await fetch(`${messageUrl}/${removableMessageId}`, requestOptions);
+                if (response.status === 200) {
+                    const newMessageList = messageList.filter(message => message._id !== removableMessageId);
+                    setMessageList(newMessageList)
+                    console.log('Message delete was successful.');
+                } else {
+                    throw new Error("Failed to delete message.")
+                }
             }
-        }
-        catch (error) {
-            console.log("Problem with message delete!")
+            catch (error) {
+                console.error("Problem with message delete!", error.message)
+            }
+        } else if (actualClientData !== undefined) {
+            try {
+                const requestOptions = {
+                    method: 'PUT',
+                    headers: { 'Content-Type': 'application/json' },
+                    body: JSON.stringify({ isActive: false })
+                };
+                const response = await fetch(`${messageUrl}/${removableMessageId}`, requestOptions);
+                if (response.status === 200) {
+                    const newMessageList = messageList.filter(message => message._id !== removableMessageId);
+                    setMessageList(newMessageList)
+                    console.log('Message remove was successful');
+                }
+            }
+            catch (error) {
+                console.log("Problem with message delete!")
+            }
         }
     };
 
@@ -79,7 +87,7 @@ const MessageContextProvider = ({ children }) => {
             messageList, setMessageList, newOrModifiedMessage, setNewOrModifiedMessage,
             originalMessage, setOriginalMessage, showMessageThanks, setShowMessageThanks,
             showTopMessageBox, setShowTopMessageBox,
-     //       deleteMessageFetch, removeMessageFetch
+            updateMessage, deleteMessage
         }}>
             {children}
         </MessageContext.Provider>
@@ -89,7 +97,7 @@ const MessageContextProvider = ({ children }) => {
 export const useMessageContext = () => useContext(MessageContext);
 
 //export const useMessagesFetch = () => useContext(MessageContext).messagesFetch();
-export const useDeleteMessageFetch = (removableMessageId) => useContext(MessageContext).deleteMessageFetch(removableMessageId);
-export const useRemoveMessageFetch = (removableMessageId) => useContext(MessageContext).removeMessageFetch(removableMessageId);
+//export const useDeleteMessageFetch = (removableMessageId) => useContext(MessageContext).deleteMessageFetch(removableMessageId);
+//export const useRemoveMessageFetch = (removableMessageId) => useContext(MessageContext).removeMessageFetch(removableMessageId);
 
 export default MessageContextProvider
